requirements.txt
Pillow
python-docx
python-whois
PyMuPDF
